requirements/local.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements/local.txt requirements/local.in
#
alabaster==0.7.12         # via sphinx
apipkg==1.5               # via execnet
asn1crypto==0.24.0        # via cryptography
astroid==2.3.1            # via pylint, pylint-celery
atomicwrites==1.3.0       # via pytest
attrs==19.2.0             # via pytest
babel==2.7.0              # via sphinx
beautifulsoup4==4.8.0
cached-property==1.5.1    # via docker-compose
cffi==1.12.3              # via cryptography
click-log==0.3.2          # via edx-lint
click==7.0                # via click-log, edx-lint
coreapi==2.3.3            # via django-rest-swagger, openapi-codec
coreschema==0.0.4         # via coreapi
coverage==4.5.4
cryptography==2.7         # via pyopenssl
ddt==1.2.1
defusedxml==0.6.0         # via djangorestframework-xml, python3-openid, social-auth-core
django-admin-sortable2==0.7.3
django-appconf==1.0.3     # via django-compressor
django-autocomplete-light==3.1.8
django-choices==1.7.0
django-compressor==2.3
django-contrib-comments==1.9.1
django-cors-headers==2.5.3
django-debug-toolbar==1.11
django-elasticsearch-debug-toolbar==2.0.0
django-extensions==2.2.1
django-filter==2.2.0
django-fsm==2.6.1
django-guardian==1.5.1
django-haystack==2.8.1
django-libsass==0.7
django-parler==2.0
django-rest-swagger==2.2.0
django-simple-history==2.7.3
django-solo==1.1.3
django-sortedm2m==2.0.0
django-stdimage==3.2.0
django-storages==1.7.2
django-taggit-autosuggest==0.3.8
django-taggit-serializer==0.1.7
django-taggit==1.1.0
django-waffle==0.17.0
django-webpack-loader==0.6.0
django==1.11.25
djangorestframework-csv==2.1.0
djangorestframework-jwt==1.11.0
djangorestframework-xml==1.4.0
djangorestframework==3.9.4
docker-compose==1.23.2
docker==3.6.0             # via docker-compose
dockerpty==0.4.1          # via docker-compose
docopt==0.6.2             # via docker-compose
docutils==0.15.2          # via sphinx
drf-dynamic-fields==0.3.1
drf-extensions==0.5.0
drf-haystack==1.8.2
dry-rest-permissions==0.1.10
edx-analytics-data-api-client==0.15.3
edx-auth-backends==2.0.2
edx-ccx-keys==1.0.0
edx-django-release-util==0.3.1
edx-django-sites-extensions==2.4.0
edx-django-utils==2.0.0
edx-drf-extensions==2.4.0
edx-i18n-tools==0.4.8
edx-lint==1.4.0
edx-opaque-keys==2.0.0
edx-rest-api-client==1.9.2
edx-sphinx-theme==1.5.0
elasticsearch==1.9.0
execnet==1.7.1            # via pytest-xdist
factory-boy==2.12.0
faker==2.0.2              # via factory-boy
freezegun==0.3.12
future==0.17.1            # via pyjwkest
html2text==2019.9.26
imagesize==1.1.0          # via sphinx
importlib-metadata==0.23  # via path.py, pluggy, pytest
isort==4.3.21
itypes==1.1.0             # via coreapi
jinja2==2.10.1            # via coreschema, sphinx
jsonfield==2.0.2
jsonschema==2.6.0         # via docker-compose
lazy-object-proxy==1.4.2  # via astroid
libsass==0.19.2           # via django-libsass
lxml==4.4.1
markdown==3.1.1
markupsafe==1.1.1         # via jinja2
mccabe==0.6.1             # via pylint
mock==3.0.5
more-itertools==7.2.0     # via pytest, zipp
ndg-httpsclient==0.5.1    # via requests
newrelic==5.0.2.126       # via edx-django-utils
oauthlib==3.1.0           # via requests-oauthlib, social-auth-core
openapi-codec==1.3.2      # via django-rest-swagger
packaging==19.2           # via pytest, sphinx
path.py==12.0.1           # via edx-i18n-tools
pathlib2==2.3.5           # via pytest
pbr==5.4.3                # via stevedore
pillow==6.1.0
pluggy==0.13.0            # via pytest
polib==1.1.0              # via edx-i18n-tools
progressbar2==3.47.0      # via django-stdimage
psutil==1.2.1             # via edx-django-utils, edx-drf-extensions
py==1.8.0                 # via pytest
pyasn1==0.4.7             # via ndg-httpsclient, requests
pycodestyle==2.5.0
pycountry==19.8.18
pycparser==2.19           # via cffi
pycryptodomex==3.9.0      # via pyjwkest
pygments==2.4.2           # via sphinx
pyinotify==0.9.6
pyjwkest==1.3.2           # via edx-drf-extensions, social-auth-core
pyjwt==1.7.1              # via djangorestframework-jwt, edx-auth-backends, edx-rest-api-client, social-auth-core
pylint-celery==0.3        # via edx-lint
pylint-django==2.0.11     # via edx-lint
pylint-plugin-utils==0.6  # via pylint-celery, pylint-django
pylint==2.4.2             # via edx-lint, pylint-celery, pylint-django, pylint-plugin-utils
pymongo==3.9.0            # via edx-opaque-keys
pyopenssl==19.0.0         # via ndg-httpsclient, requests
pyparsing==2.4.2          # via packaging
pytest-cov==2.7.1
pytest-django-ordering==1.2.0
pytest-django==3.5.1
pytest-forked==1.0.2      # via pytest-xdist
pytest-responses==0.4.0
pytest-xdist==1.30.0
pytest==5.2.0
python-dateutil==2.8.0
python-utils==2.3.0       # via progressbar2
python3-openid==3.1.0     # via social-auth-core
pytz==2019.2
pyyaml==3.13              # via docker-compose, edx-django-release-util, edx-i18n-tools
rcssmin==1.0.6            # via django-compressor
requests-oauthlib==1.2.0  # via social-auth-core
requests[security]==2.11.1
responses==0.10.6
rest-condition==1.0.3     # via edx-drf-extensions
rjsmin==1.1.0             # via django-compressor
selenium==3.141.0
semantic-version==2.8.2   # via edx-drf-extensions
simple-salesforce==0.74.3
simplejson==3.16.0        # via django-rest-swagger
six==1.12.0               # via astroid, cryptography, django-appconf, django-extensions, django-simple-history, django-sortedm2m, django-taggit-serializer, djangorestframework-csv, docker-compose, dockerpty, edx-auth-backends, edx-ccx-keys, edx-django-release-util, edx-drf-extensions, edx-i18n-tools, edx-lint, edx-opaque-keys, edx-sphinx-theme, faker, freezegun, libsass, mock, packaging, pathlib2, progressbar2, pyjwkest, pyopenssl, pytest-xdist, python-dateutil, python-utils, responses, social-auth-app-django, social-auth-core, stevedore, transifex-client, unicode-slugify, websocket-client
slumber==0.7.1            # via edx-rest-api-client
snowballstemmer==1.9.1    # via sphinx
social-auth-app-django==1.2.0  # via edx-auth-backends
social-auth-core[openidconnect]==1.7.0  # via edx-auth-backends, social-auth-app-django
soupsieve==1.9.4          # via beautifulsoup4
sphinx==2.2.0
sphinxcontrib-applehelp==1.0.1  # via sphinx
sphinxcontrib-devhelp==1.0.1  # via sphinx
sphinxcontrib-htmlhelp==1.0.2  # via sphinx
sphinxcontrib-jsmath==1.0.1  # via sphinx
sphinxcontrib-qthelp==1.0.2  # via sphinx
sphinxcontrib-serializinghtml==1.1.3  # via sphinx
sqlparse==0.3.0           # via django-debug-toolbar
stevedore==1.31.0         # via edx-opaque-keys
testfixtures==6.10.0
text-unidecode==1.3       # via faker
texttable==0.9.1          # via docker-compose
transifex-client==0.12.5
typed-ast==1.4.0          # via astroid
unicode-slugify==0.1.3
unicodecsv==0.14.1        # via djangorestframework-csv
unidecode==1.1.1          # via unicode-slugify
uritemplate==3.0.0        # via coreapi
urllib3==1.25.6           # via elasticsearch, selenium, transifex-client
wcwidth==0.1.7            # via pytest
websocket-client==0.56.0  # via docker-compose
wrapt==1.11.2             # via astroid
xss-utils==0.1.1
zipp==0.6.0               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools==41.2.0        # via markdown, sphinx
